Moved country creation modal into a separate component
